size == unsigned long
git-svn-id: file:///source/coreland/chrono/trunk@2956 6e7d7e42-f3d6-da11-80d8-00d0b73f4485

+++ leaps_read.c
@@ -33,7 +33,7 @@ int leapsecs_read(void)
   if (!mem) goto FAIL;
 
   size = read(fd, mem, sb.st_size);
-  if (size != sb.st_size) { dealloc(mem); goto FAIL; }
+  if (size != (unsigned long) sb.st_size) { dealloc(mem); goto FAIL; }
 
   size = sb.st_size / sizeof(struct tai);
   for (ind = 0; ind < size; ++ind) {
